k8s: Make it work even if kubectl is not installed on the machine
There's code in the k8s provider that takes care of pulling the
kubectl binary from the cluster and use that one for subsequent
operations; however, right before that happens, a few settings
are changed in kubeconfig, which means that it's necessary for
the host system to provide its own kubectl binary.

Invert the order of these operations, and make sure that we
always invoke the binary obtained from the cluster. This way, it
is possible to successfully bring up a cluster on a machine that
doesn't have kubectl installed.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>